+++ agents/exec/che-exec-agent-shared/src/main/java/org/eclipse/che/agent/exec/shared/dto/ProcessStartResponseDto.java
@@ -8,10 +8,9 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.ide.api.command.exec.dto;
+package org.eclipse.che.agent.exec.shared.dto;
 
 import org.eclipse.che.dto.shared.DTO;
-import org.eclipse.che.ide.api.command.exec.dto.event.DtoWithPid;
 
 @DTO
 public interface ProcessStartResponseDto extends DtoWithPid {
