+++ agents/exec/che-exec-agent-shared/src/main/java/org/eclipse/che/agent/exec/shared/dto/ProcessUnSubscribeRequestDto.java
@@ -8,16 +8,19 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.ide.api.command.exec.dto;
+package org.eclipse.che.agent.exec.shared.dto;
 
 import org.eclipse.che.dto.shared.DTO;
-import org.eclipse.che.ide.api.command.exec.dto.event.DtoWithPid;
 
 @DTO
-public interface UpdateSubscriptionRequestDto extends DtoWithPid {
-  UpdateSubscriptionRequestDto withPid(int pid);
+public interface ProcessUnSubscribeRequestDto extends DtoWithPid {
+  ProcessUnSubscribeRequestDto withPid(int pid);
 
   String getEventTypes();
 
-  UpdateSubscriptionRequestDto withEventTypes(String eventTypes);
+  ProcessUnSubscribeRequestDto withEventTypes(String eventTypes);
+
+  String getAfter();
+
+  ProcessUnSubscribeRequestDto withAfter(String after);
 }
