+++ agents/exec/che-exec-agent-shared/src/main/java/org/eclipse/che/agent/exec/shared/dto/ProcessUnSubscribeResponseDto.java
@@ -8,20 +8,15 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.ide.api.command.exec.dto;
+package org.eclipse.che.agent.exec.shared.dto;
 
 import org.eclipse.che.dto.shared.DTO;
-import org.eclipse.che.ide.api.command.exec.dto.event.DtoWithPid;
 
 @DTO
-public interface UpdateSubscriptionResponseDto extends DtoWithPid {
-  UpdateSubscriptionResponseDto withPid(int pid);
-
-  String getEventTypes();
-
-  UpdateSubscriptionResponseDto withEventTypes(String eventTypes);
+public interface ProcessUnSubscribeResponseDto extends DtoWithPid {
+  ProcessUnSubscribeResponseDto withPid(int pid);
 
   String getText();
 
-  UpdateSubscriptionResponseDto withText(String text);
+  ProcessUnSubscribeResponseDto withText(String text);
 }
