+++ agents/exec/che-exec-agent-shared/src/main/java/org/eclipse/che/agent/exec/shared/dto/event/ProcessStdOutEventDto.java
@@ -8,12 +8,20 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.ide.api.command.exec.dto;
+package org.eclipse.che.agent.exec.shared.dto.event;
 
+import org.eclipse.che.agent.exec.shared.dto.DtoWithPid;
 import org.eclipse.che.dto.shared.DTO;
-import org.eclipse.che.ide.api.command.exec.dto.event.DtoWithPid;
 
 @DTO
-public interface GetProcessRequestDto extends DtoWithPid {
-  GetProcessRequestDto withPid(int pid);
+public interface ProcessStdOutEventDto extends DtoWithPid {
+  ProcessStdOutEventDto withPid(int pid);
+
+  String getTime();
+
+  ProcessStdOutEventDto withTime(String time);
+
+  String getText();
+
+  ProcessStdOutEventDto withText(String text);
 }
