+++ agents/exec/installer/src/assembly/assembly.xml
