+++ agents/exec/installer/src/main/resources/installers/1.0.0/org.eclipse.che.exec.json
