+++ ide/che-core-ide-api/pom.xml
@@ -52,6 +52,15 @@
             <artifactId>validation-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.eclipse.che</groupId>
+            <artifactId>che-exec-agent-shared</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che</groupId>
+            <artifactId>che-exec-agent-shared</artifactId>
+            <classifier>sources</classifier>
+        </dependency>
+        <dependency>
             <groupId>org.eclipse.che.core</groupId>
             <artifactId>che-core-api-auth-shared</artifactId>
         </dependency>
@@ -288,7 +297,7 @@
                         <package>org.eclipse.che.api.user</package>
                         <package>org.eclipse.che.api.debug.shared.dto</package>
                         <package>org.eclipse.che.ide.api.debug.dto</package>
-                        <package>org.eclipse.che.ide.api.command.exec.dto</package>
+                        <package>org.eclipse.che.agent.exec.shared.dto</package>
                     </dtoPackages>
                     <outputDirectory>${dto-generator-out-directory}</outputDirectory>
                     <genClassName>org.eclipse.che.ide.api.dto.DtoClientImpls</genClassName>
@@ -330,6 +339,7 @@
                                 <dep>org.eclipse.che.core:che-core-api-debug-shared</dep>
                                 <dep>org.eclipse.che.core:che-core-api-ssh-shared</dep>
                                 <dep>org.eclipse.che.core:che-core-gwt-logger</dep>
+                                <dep>org.eclipse.che:che-exec-agent-shared</dep>
                             </ignoredUnusedDeclaredDependencies>
                         </configuration>
                     </execution>
