+++ ide/che-core-ide-api/src/main/java/org/eclipse/che/ide/api/command/exec/ProcessStartedEventHandler.java
@@ -12,8 +12,8 @@ package org.eclipse.che.ide.api.command.exec;
 
 import com.google.inject.Inject;
 import javax.inject.Singleton;
+import org.eclipse.che.agent.exec.shared.dto.event.ProcessStartedEventDto;
 import org.eclipse.che.api.core.jsonrpc.commons.RequestHandlerConfigurator;
-import org.eclipse.che.ide.api.command.exec.dto.event.ProcessStartedEventDto;
 import org.eclipse.che.ide.util.loging.Log;
 
 /**
