+++ ide/che-core-ide-app/pom.xml
@@ -61,6 +61,15 @@
             <artifactId>validation-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.eclipse.che</groupId>
+            <artifactId>che-exec-agent-shared</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che</groupId>
+            <artifactId>che-exec-agent-shared</artifactId>
+            <classifier>sources</classifier>
+        </dependency>
+        <dependency>
             <groupId>org.eclipse.che.core</groupId>
             <artifactId>che-core-api-auth-shared</artifactId>
         </dependency>
@@ -390,6 +399,7 @@
                                 <dep>org.eclipse.che.core:che-core-api-user-shared</dep>
                                 <dep>org.eclipse.che.core:che-core-api-project-templates-shared</dep>
                                 <dep>org.eclipse.che.core:che-core-api-debug-shared</dep>
+                                <dep>org.eclipse.che:che-exec-agent-shared</dep>
                             </ignoredUnusedDeclaredDependencies>
                         </configuration>
                     </execution>
