+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/command/execute/JsonRpcExecAgentCommandManager.java
@@ -16,33 +16,33 @@ import com.google.inject.Inject;
 import java.util.List;
 import java.util.function.Consumer;
 import javax.inject.Singleton;
+import org.eclipse.che.agent.exec.shared.dto.DtoWithPid;
+import org.eclipse.che.agent.exec.shared.dto.GetProcessLogsRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.GetProcessLogsResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.GetProcessRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.GetProcessResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.GetProcessesRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.GetProcessesResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessKillRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessKillResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessStartRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessStartResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessSubscribeRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessSubscribeResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessUnSubscribeRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.ProcessUnSubscribeResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.UpdateSubscriptionRequestDto;
+import org.eclipse.che.agent.exec.shared.dto.UpdateSubscriptionResponseDto;
+import org.eclipse.che.agent.exec.shared.dto.event.ProcessDiedEventDto;
+import org.eclipse.che.agent.exec.shared.dto.event.ProcessStartedEventDto;
+import org.eclipse.che.agent.exec.shared.dto.event.ProcessStdErrEventDto;
+import org.eclipse.che.agent.exec.shared.dto.event.ProcessStdOutEventDto;
 import org.eclipse.che.api.core.jsonrpc.commons.JsonRpcPromise;
 import org.eclipse.che.api.core.jsonrpc.commons.RequestTransmitter;
 import org.eclipse.che.api.core.model.workspace.config.Command;
 import org.eclipse.che.ide.api.command.exec.ExecAgentCommandManager;
 import org.eclipse.che.ide.api.command.exec.ExecAgentConsumer;
 import org.eclipse.che.ide.api.command.exec.ExecAgentEventManager;
-import org.eclipse.che.ide.api.command.exec.dto.GetProcessLogsRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.GetProcessLogsResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.GetProcessRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.GetProcessResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.GetProcessesRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.GetProcessesResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessKillRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessKillResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessStartRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessStartResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessSubscribeRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessSubscribeResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessUnSubscribeRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.ProcessUnSubscribeResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.UpdateSubscriptionRequestDto;
-import org.eclipse.che.ide.api.command.exec.dto.UpdateSubscriptionResponseDto;
-import org.eclipse.che.ide.api.command.exec.dto.event.DtoWithPid;
-import org.eclipse.che.ide.api.command.exec.dto.event.ProcessDiedEventDto;
-import org.eclipse.che.ide.api.command.exec.dto.event.ProcessStartedEventDto;
-import org.eclipse.che.ide.api.command.exec.dto.event.ProcessStdErrEventDto;
-import org.eclipse.che.ide.api.command.exec.dto.event.ProcessStdOutEventDto;
 import org.eclipse.che.ide.dto.DtoFactory;
 import org.eclipse.che.ide.util.loging.Log;
