+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/machine/ProcessesOutputRestorer.java
@@ -16,9 +16,9 @@ import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.google.web.bindery.event.shared.EventBus;
 import java.util.function.Consumer;
+import org.eclipse.che.agent.exec.shared.dto.DtoWithPid;
 import org.eclipse.che.ide.api.app.AppContext;
 import org.eclipse.che.ide.api.command.exec.ExecAgentCommandManager;
-import org.eclipse.che.ide.api.command.exec.dto.event.DtoWithPid;
 import org.eclipse.che.ide.api.workspace.event.ExecAgentServerRunningEvent;
 import org.eclipse.che.ide.api.workspace.model.RuntimeImpl;
 import org.eclipse.che.ide.api.workspace.model.WorkspaceImpl;
