+++ plugins/plugin-machine/che-plugin-machine-ext-server/src/main/java/org/eclipse/che/ide/ext/machine/server/MachineModule.java
@@ -13,15 +13,14 @@ package org.eclipse.che.ide.ext.machine.server;
 import com.google.inject.AbstractModule;
 import org.eclipse.che.ide.ext.machine.server.ssh.KeysInjector;
 import org.eclipse.che.ide.ext.machine.server.ssh.WorkspaceSshKeys;
+import org.eclipse.che.inject.DynaModule;
 
-// @DynaModule
+@DynaModule
 public class MachineModule extends AbstractModule {
 
   @Override
   protected void configure() {
     bind(KeysInjector.class).asEagerSingleton();
     bind(WorkspaceSshKeys.class).asEagerSingleton();
-    // FIXME: spi
-    //        bind(RecipeScriptDownloadService.class);
   }
 }
